[Fix-203] Implementando nova estrutura de pastas para modulo home.
