+++ lib/src/app/modules/configuration/contact/contact_module.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:is_it_safe_app/src/app/modules/configuration/contact/presenter/bloc/contact_bloc.dart';
 import 'package:is_it_safe_app/src/app/modules/configuration/contact/presenter/pages/contact_page.dart';
-import 'package:is_it_safe_app/src/domain/use_case/get_user_name_use_case.dart';
+import 'package:is_it_safe_app/src/app/modules/home/domain/usecases/get_user_name_use_case.dart';
 
 class ContactModule extends Module {
   @override
