+++ lib/src/app/modules/configuration/contact/presenter/bloc/contact_bloc.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:is_it_safe_app/generated/l10n.dart';
 import 'package:is_it_safe_app/src/core/interfaces/safe_bloc.dart';
-import 'package:is_it_safe_app/src/domain/use_case/get_user_name_use_case.dart';
+import 'package:is_it_safe_app/src/app/modules/home/domain/usecases/get_user_name_use_case.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class ContactBloc extends SafeBloC {
