+++ lib/src/app/modules/home/presenter/VOs/home_drawer_vo.dart
@@ -1,0 +1,9 @@
+class HomeDrawerVO {
+  String userName;
+  String userImage;
+
+  HomeDrawerVO({
+    required this.userName,
+    required this.userImage,
+  });
+}
