+++ lib/src/app/modules/home/presenter/widgets/home_drawer.dart
@@ -161,13 +161,3 @@ class DrawerItem extends StatelessWidget {
     );
   }
 }
-
-class HomeDrawerVO {
-  String userName;
-  String userImage;
-
-  HomeDrawerVO({
-    required this.userName,
-    required this.userImage,
-  });
-}
