+++ lib/src/app/modules/location/domain/usecases/get_locations_by_name_use_case.dart
@@ -3,7 +3,7 @@ import 'package:is_it_safe_app/src/app/modules/location/domain/entities/location
 import 'package:is_it_safe_app/src/service/api/modules/search/search_service_interface.dart';
 import 'package:result_dart/result_dart.dart';
 
-import '../../app/modules/search/error/safe_search_error.dart';
+import '../../../search/error/safe_search_error.dart';
 
 class GetLocationsByNameUseCase extends SafeUseCase {
   final ISearchService service;
