+++ test/app/modules/location_module/domain/save_location_use_case_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:is_it_safe_app/src/app/modules/location/domain/entities/request/request_save_location.dart';
-import 'package:is_it_safe_app/src/domain/use_case/save_location_use_case.dart';
+import 'package:is_it_safe_app/src/app/modules/location/domain/usecases/save_location_use_case.dart';
 import 'package:modular_test/modular_test.dart';
 
 import '../mocks/mock_location_service.dart';
